move jobq to pkg
